backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.27.0
pydantic==2.5.0  # ATENCAO: wheels oficiais ate Python 3.12; Py3.13+ exige Rust/Cargo para pydantic-core [ASSUNCAO]

# Database (F5 Histórico)
psycopg2-binary==2.9.10  # bumped para evitar build local (pg_config) [ASSUNCAO]
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# Alertas (F6)
celery==5.3.4
redis==5.0.1
pyyaml==6.0.1
